feat: Room 참가, 생성 로직 추가 구현
